[clean] modalをdialogに変更 #92
